refactor: extract hook_registry.py from __init__.py
Move catch/defeat hook system (~45 lines) into its own module.
CatchPokemonHook, DefeatPokemonHook, and the hook registration
functions now live in hook_registry.py, importing singletons directly.
